Card Model Functionality
This class is a representation of a single card in the game Set. This
commit includes a build out of all the attributes that are needed to
model a card; as such it has four properties (Color, Fill, Shape,
Number). These properties can only be changed at the instantiation of an
instance of a class, while there are access methods the properties can
not be changed once the card is created.