Deck Model Functionality
This class is a representation of a deck in the game of Set. This commit
includes a build out of all the attributes and some functionality of a
Set deck. This class will return a card at a specific index (passed in),
the next card at the logical top of the deck, the next three cards from
the logical top of the deck, and it will shuffle the deck (shuffle
algorithm is that in the java.util.Collections class). In the Code there
are commented out questions that I have, a later commit will remove
these.