Match Model Funtionality
This class Is a representation of Match in the game set. This commit
includes a built out of all the attributes of a match. As such, this
class takes an array of cards (which should only ever contain three
cards). You can change the set given at instantiation, and check to see
if the passed array of cards is a match.