PlayField Model Functionality
This class Is a representation of the Playing Field in the game Set.
This commit includes a build out of all the attributes and functionality
that concern the playing field. As such, you can analyze the deck (see
if there are any matches), you can get those matches, and you can get a
history of the board, as well as update the field (the newField()
operates on immutability. You must create a new version of the field and
pass it in).